refactor: travel_calendar.py 가 너무 많은 역할과 크기를 가지고 있어서 목적에 맞게 분리
